requirements.txt
django==3.1.5
requests==2.25.1
psycopg2==2.8.6
environs==9.3.0
djangorestframework==3.12.2
responses==0.12.1
